Refactor: Integrate SR optimization into MarketAnalysisSubPipeline
