+++ src/training/steps/market_analysis/REFACTORING_SUMMARY.md
@@ -56,15 +56,16 @@ The `step02_5_sr_optimization.py` file has been successfully moved from `data_pr
   - Pipeline state management
   - Detailed logging and monitoring
 
-### Backward Compatibility Wrapper (`sr_optimization_compatibility.py`)
-- **Purpose**: Provides full backward compatibility with original SROptimizationStep interface
-- **Class**: `SROptimizationStep`
+### Backward Compatibility Integration
+- **Purpose**: Direct integration of backward compatibility into MarketAnalysisSubPipeline
+- **Class**: `MarketAnalysisSubPipeline` (aliased as `SROptimizationStep`)
 - **Key Features**:
   - Identical interface to original SROptimizationStep
   - `execute(training_input, pipeline_state)` method maintained
-  - Automatic configuration conversion
+  - Automatic configuration conversion from old dict format
   - Sequential execution of all three SR stages
   - Comprehensive error handling and logging
+  - No separate wrapper file needed
 
 ## Import Updates
 
@@ -76,9 +77,9 @@ The `step02_5_sr_optimization.py` file has been successfully moved from `data_pr
 
 ### Import Path Changes
 - **Old**: `src.training.steps.data_collection.data_preparation.step02_5_sr_optimization`
-- **New**: `src.training.steps.market_analysis.sr_optimization_compatibility` (SROptimizationStep)
-- **Backward Compatibility**: Full interface compatibility maintained
-- **Factory Integration**: All factory files updated to use compatibility wrapper
+- **New**: `src.training.steps.market_analysis.sub_pipeline` (MarketAnalysisSubPipeline)
+- **Backward Compatibility**: Full interface compatibility maintained via direct integration
+- **Factory Integration**: All factory files updated to use sub_pipeline directly
 
 ## Benefits of Refactoring
 
@@ -124,7 +125,7 @@ ml_step = SRMLLearningStep(config)
 # Original interface (fully backward compatible)
 from src.training.steps.market_analysis import SROptimizationStep
 
-sr_step = SROptimizationStep(config)
+sr_step = SROptimizationStep(config)  # Same as MarketAnalysisSubPipeline
 result = await sr_step.execute(training_input, pipeline_state)
 
 # Or use the sub_pipeline directly
@@ -145,7 +146,8 @@ The refactoring maintains **100% backward compatibility**:
 - ✅ Configuration parameters are automatically converted and preserved
 - ✅ Return values and execution flow are identical
 - ✅ All factory and validation systems updated seamlessly
-- ✅ Comprehensive compatibility wrapper ensures zero breaking changes
+- ✅ Direct integration into MarketAnalysisSubPipeline ensures zero breaking changes
+- ✅ No separate wrapper file needed - cleaner architecture
 
 ## Verification
 
@@ -154,9 +156,10 @@ The refactoring maintains **100% backward compatibility**:
 - ✅ Old files removed from original locations
 - ✅ Package structure properly initialized
 - ✅ **100% backward compatibility maintained**
-- ✅ Compatibility wrapper tested and verified
+- ✅ Direct integration into MarketAnalysisSubPipeline verified
 - ✅ All factory systems updated and working
 - ✅ Import structure validated across all modules
+- ✅ No separate wrapper file needed - cleaner architecture
 
 ## Next Steps
