+++ src/training/steps/market_analysis/__init__.py
@@ -4,12 +4,14 @@ from .sr_detection import SRDetectionStep
 from .sr_clustering import SRClusteringStep
 from .sr_ml_learning import SRMLLearningStep
 from .sub_pipeline import MarketAnalysisSubPipeline
-from .sr_optimization_compatibility import SROptimizationStep
+
+# Backward compatibility alias
+SROptimizationStep = MarketAnalysisSubPipeline
 
 __all__ = [
     'SRDetectionStep', 
     'SRClusteringStep',
     'SRMLLearningStep',
     'MarketAnalysisSubPipeline',
-    'SROptimizationStep'  # Backward compatibility wrapper
+    'SROptimizationStep'  # Backward compatibility alias
 ]
