+++ src/utils/enhanced_step_wrapper.py
@@ -189,7 +189,7 @@ class EnhancedPipelineManager:
     def enhance_all_pipeline_steps(self) -> Dict[str, Type]:
         """Enhance all known pipeline steps."""
         enhanced_steps = {}
-        step_configs = {'step02_data_reading': ('src.training.steps.data_collection.step02_data_reading', 'DataReadingStep'), 'step2_5_sr_optimization': ('src.training.steps.market_analysis.sr_optimization_compatibility', 'SROptimizationStep'), 'step03_hmm_regime_discovery': ('src.training.steps.data_collection.data_preparation.step03_hmm_regime_discovery', 'Step03HMMRegimeDiscovery'), 'step04_regime_data_splitting': ('src.training.steps.step08_regime_data_splitting', 'RegimeDataSplittingStep'), 'step05_labeling': ('src.training.steps.step5_labeling', 'LabelingStep'), 'step06_advanced_feature_engineering': ('src.training.steps.data_collection.feature_engineering.step06_feature_engineering', 'FeatureEngineeringStep'), 'step08_advanced_feature_selection': ('src.training.steps.data_collection.feature_engineering.step08_advanced_feature_selection', 'FeatureSelectionStep'), 'step09_hmm_based_training': ('src.training.steps.model_training.step09_hmm_based_training', 'HMMTrainingStep'), 'step10_unified_regime_intelligence': ('src.training.steps.model_training.step10_unified_regime_intelligence', 'RegimeIntelligenceStep'), 'step11_analyst_creation': ('src.training.steps.model_training.step11_analyst_creation', 'AnalystCreationStep'), 'step12_analyst_enhancement': ('src.training.steps.model_training.step12_analyst_enhancement', 'AnalystEnhancementStep'), 'step13_analyst_ensemble_creation': ('src.training.steps.model_training.step13_analyst_ensemble_creation', 'AnalystEnsembleStep'), 'step14_tactician_labeling': ('src.training.steps.model_training.step14_tactician_labeling', 'TacticianLabelingStep'), 'step15_tactician_specialist_training': ('src.training.steps.model_training.step15_tactician_specialist_training', 'TacticianTrainingStep'), 'step16_confidence_calibration': ('src.training.steps.model_training.validation.step16_confidence_calibration', 'ConfidenceCalibrationStep'), 'step17_parameter_optimization': ('src.training.steps.optimisation.step17_final_parameters_optimization', 'ParameterOptimizationStep'), 'step18_walk_forward_validation': ('src.training.steps.model_training.validation.step18_walk_forward_validation', 'WalkForwardValidationStep'), 'step19_monte_carlo_validation': ('src.training.steps.model_training.validation.step19_monte_carlo_validation', 'MonteCarloValidationStep'), 'step20_ab_testing': ('src.training.steps.model_training.validation.step20_ab_testing', 'ABTestingStep'), 'step21_model_persistence': ('src.training.steps.backtesting.step21_saving', 'ModelPersistenceStep')}
+        step_configs = {'step02_data_reading': ('src.training.steps.data_collection.step02_data_reading', 'DataReadingStep'), 'step2_5_sr_optimization': ('src.training.steps.market_analysis.sub_pipeline', 'MarketAnalysisSubPipeline'), 'step03_hmm_regime_discovery': ('src.training.steps.data_collection.data_preparation.step03_hmm_regime_discovery', 'Step03HMMRegimeDiscovery'), 'step04_regime_data_splitting': ('src.training.steps.step08_regime_data_splitting', 'RegimeDataSplittingStep'), 'step05_labeling': ('src.training.steps.step5_labeling', 'LabelingStep'), 'step06_advanced_feature_engineering': ('src.training.steps.data_collection.feature_engineering.step06_feature_engineering', 'FeatureEngineeringStep'), 'step08_advanced_feature_selection': ('src.training.steps.data_collection.feature_engineering.step08_advanced_feature_selection', 'FeatureSelectionStep'), 'step09_hmm_based_training': ('src.training.steps.model_training.step09_hmm_based_training', 'HMMTrainingStep'), 'step10_unified_regime_intelligence': ('src.training.steps.model_training.step10_unified_regime_intelligence', 'RegimeIntelligenceStep'), 'step11_analyst_creation': ('src.training.steps.model_training.step11_analyst_creation', 'AnalystCreationStep'), 'step12_analyst_enhancement': ('src.training.steps.model_training.step12_analyst_enhancement', 'AnalystEnhancementStep'), 'step13_analyst_ensemble_creation': ('src.training.steps.model_training.step13_analyst_ensemble_creation', 'AnalystEnsembleStep'), 'step14_tactician_labeling': ('src.training.steps.model_training.step14_tactician_labeling', 'TacticianLabelingStep'), 'step15_tactician_specialist_training': ('src.training.steps.model_training.step15_tactician_specialist_training', 'TacticianTrainingStep'), 'step16_confidence_calibration': ('src.training.steps.model_training.validation.step16_confidence_calibration', 'ConfidenceCalibrationStep'), 'step17_parameter_optimization': ('src.training.steps.optimisation.step17_final_parameters_optimization', 'ParameterOptimizationStep'), 'step18_walk_forward_validation': ('src.training.steps.model_training.validation.step18_walk_forward_validation', 'WalkForwardValidationStep'), 'step19_monte_carlo_validation': ('src.training.steps.model_training.validation.step19_monte_carlo_validation', 'MonteCarloValidationStep'), 'step20_ab_testing': ('src.training.steps.model_training.validation.step20_ab_testing', 'ABTestingStep'), 'step21_model_persistence': ('src.training.steps.backtesting.step21_saving', 'ModelPersistenceStep')}
         for step_name, (module_path, class_name) in step_configs.items():
             try:
                 module = __import__(module_path, fromlist=[class_name])
